requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
pyrebase4>=4.0.0
pyrebase4>=4.0.0
